osdmaptool: Additions for OBJECT-1899
- mark_up/mark_in flags: Allow to mark up/in OSD that are down/out.
Useful to know how much data will move if we recreate an OSD.
OBJECT-1881 is a result of this patch.
- fix mark_up_in which makes all the down/out OSD in the cluster up/in.
Never had to use this but it sets the crush weight to 1.0 even if the
OSDs already had a crush weight. That'd mess up a lot of things if we
were to use it
- adjust-crush-weight flag: Allows to change the weight of OSD(s)
directly via osdmaptool. Previously we had to do it manually (see
Simulate PG movement before CRUSHmap changes in confluence). Usage:
 --adjust-crush-weight <osdid>:<new_weight>,<osdid>:<new_weight>,...

Signed-off-by: Alex Marangone <amarangone@digitalocean.com>